+++ src/rendering/vcPolygonModel.h
@@ -51,7 +51,7 @@ struct vcPolygonModel
 udResult vcPolygonModel_CreateShaders();
 udResult vcPolygonModel_DestroyShaders();
 
-udResult vcPolygonModel_CreateFromRawVertexData(vcPolygonModel **ppPolygonModel, const void *pVerts, const uint16_t vertCount, const vcVertexLayoutTypes *pMeshLayout, const int totalTypes, const uint16_t *pIndices = nullptr, const uint16_t indexCount = 0);
+udResult vcPolygonModel_CreateFromRawVertexData(vcPolygonModel **ppPolygonModel, const void *pVerts, const uint32_t vertCount, const vcVertexLayoutTypes *pMeshLayout, const int totalTypes, const uint32_t *pIndices = nullptr, const uint32_t indexCount = 0);
 udResult vcPolygonModel_CreateFromURL(vcPolygonModel **ppModel, const char *pURL, udWorkerPool *pWorkerPool);
 
 udResult vcPolygonModel_Destroy(vcPolygonModel **ppModel);
